+++ crates/cli/src/session.rs
@@ -19,6 +19,7 @@ use starbase_console::Console;
 use starbase_console::ui::{OwnedOrShared, Progress, ProgressDisplay, ProgressReporter};
 use starbase_utils::envx;
 use std::collections::VecDeque;
+use std::env;
 use std::sync::Arc;
 use tokio::task::JoinSet;
 use tracing::{debug, instrument};
@@ -333,6 +334,13 @@ impl ProtoSession {
         ProgressInstance { reporter, handle }
     }
 
+    pub fn is_tool_exec_command(&self) -> bool {
+        matches!(
+            self.cli.command,
+            Commands::Exec(_) | Commands::Run(_) | Commands::Shell(_)
+        )
+    }
+
     pub fn is_json_format(&self) -> bool {
         self.console.is_json_format()
     }
@@ -351,9 +359,11 @@ impl AppSession for ProtoSession {
     type Error = miette::Report;
 
     async fn startup(&mut self) -> AppResult<Self::Error> {
-        if ai_env::is_ai_agent()
+        // Do not show for run/exec commands, as we don't want to pollute the output
+        if !self.is_tool_exec_command()
             && self.cli.stdout_owner() == StdoutOwner::Reporter
             && self.cli.reporter_format() == ReporterFormat::Ndjson
+            && ai_env::is_ai_agent()
         {
             self.console.message("Detected an AI agent environment, printing as NDJSON. Trace logs are written to stderr, while user-facing logs are written to stdout.")?;
         }